+++ src/config.rs
@@ -34,20 +34,20 @@ pub(crate) struct Config {
     pub(crate) language: Language,
     pub(crate) watch: Option<Watch>,
     pub(crate) resend_resource_subscriptions: Option<bool>,
-    pub(crate) build: Option<BuildConfig>,
+    pub(crate) build: Option<CommandConfig>,
+    pub(crate) run: Option<CommandConfig>,
 }
 
 #[derive(Deserialize)]
-
-pub(crate) struct BuildConfig {
+pub(crate) struct CommandConfig {
     pub(crate) command: Option<String>,
     pub(crate) args: Option<Vec<String>>,
 }
 
 #[derive(Deserialize)]
 pub(crate) struct Watch {
-    pub(crate) default_watch_paths: Option<Vec<String>>,
-    pub(crate) extra_watch_paths: Option<Vec<String>>,
+    pub(crate) default_paths: Option<Vec<String>>,
+    pub(crate) extra_paths: Option<Vec<String>>,
 }
 
 pub(crate) fn read_from_toml(path: &Path) -> Result<Config, eyre::Report> {