+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
-- removed log message from stdout
+- Removed log message from stdout.
+- Corrected starting node process in Windows.
 
 ## [0.1.0] - 2025-01-05