+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Removed log message from stdout.
 - Corrected starting node process in Windows.
+- Removed extra endline from initialize output.
 
 ## [0.1.0] - 2025-01-05