+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+WIP
+
+## [0.2.1] - 2025-01-08
+
 ### Fixed
 
 - Correctly reading configuration for default_paths and extra_paths.
@@ -53,7 +57,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `synf dev` sends list_updated for tools/prompts/resources after server restart.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/strowk/synf/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/strowk/synf/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/strowk/synf/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/strowk/synf/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/strowk/synf/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/strowk/synf/compare/v0.0.1...v0.1.0